12.04.2018: Added Role-Check on Login
Role-button only visible if User has Role "admin".
Direct URL still works, is in ToDo.
DB:Adjusted User-Table to [id, password, userName, role]
Person-Table is now [id, age, name]
Role-Table is now [id, name]
