+++ WebContent/pages/protected/templates/left.xhtml
@@ -13,7 +13,7 @@
 			<br />
 			<p:commandButton styleClass="menuButton" icon="ui-icon-arrowstop-1-e" rendered="true" action="/pages/protected/language/languageIndex.xhtml?faces-redirect=true" value="#{bundle.languagePlural}" ajax="false" immediate="true" />
 			<br />
-			<p:commandButton styleClass="menuButton" icon="ui-icon-arrowstop-1-e" rendered="#{personBean.admin}" action="/pages/protected/role/roleIndex.xhtml?faces-redirect=true" value="#{bundle.rolePlural}" ajax="false" immediate="true" />
+			<p:commandButton styleClass="menuButton" icon="ui-icon-arrowstop-1-e" rendered="#{userBean.admin}" action="/pages/protected/role/roleIndex.xhtml?faces-redirect=true" value="#{bundle.rolePlural}" ajax="false" immediate="true" />
 			<br />
 		</h:form>
 	</ui:composition>
